refactor: move filter options hooks
