+++ apps/admin-ui/src/hooks/useReservationUnitOptions.ts
@@ -1,6 +1,35 @@
 import { useEffect } from "react";
-import { useReservationUnitsFilterParamsQuery } from "@gql/gql-types";
+import { gql } from "@apollo/client";
 import { filterNonNullable } from "common/src/helpers";
+import { useReservationUnitsFilterParamsQuery } from "@gql/gql-types";
+
+export const RESERVATION_UNITS_FILTER_PARAMS_QUERY = gql`
+  query ReservationUnitsFilterParams(
+    $after: String
+    $unit: [Int]
+    $orderBy: [ReservationUnitOrderingChoices]
+  ) {
+    reservationUnits(
+      after: $after
+      onlyWithPermission: true
+      unit: $unit
+      orderBy: $orderBy
+    ) {
+      edges {
+        node {
+          id
+          nameFi
+          pk
+        }
+      }
+      pageInfo {
+        endCursor
+        hasNextPage
+      }
+      totalCount
+    }
+  }
+`;
 
 export function useReservationUnitOptions() {
   const { data, loading, fetchMore } = useReservationUnitsFilterParamsQuery();
